Fix what goes into interpreter.
Before the haskell interpreter was run with a concatenation of user provided function string and stdin string and the result should be the function applied on the input.
Standard input string was first transformed to a string resambling a haskell code for a string and later evaluated by interpreter back to a string.
One problem was that quotes in this stdin string were not escaped leading to `echo 'Hey "missiles".' | fsh 'id'` resulting in compiler error!
Second problem was the lack of brackets resulting in `fsh 'const []'` parsing but `fsh 'const $ []'` parsing differently resulting in an error!
This change fixes both problems by instead passing into the interpreter only the function string, evaluating it to haskell function and then applying it to the string.
This simplifies code also.